Exception handling - delete

+++ src/main/java/com/saadehweb/springdemo/services/UserService.java
@@ -2,8 +2,11 @@ package com.saadehweb.springdemo.services;
 
 import com.saadehweb.springdemo.entities.User;
 import com.saadehweb.springdemo.repositories.UserRepository;
+import com.saadehweb.springdemo.services.exception.DatabaseException;
 import com.saadehweb.springdemo.services.exception.ResourceNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -29,7 +32,17 @@ public class UserService {
     }
 
     public void delete(Long id){
-        repository.deleteById(id);
+        try {
+            if (!repository.existsById(id)){
+                throw new EmptyResultDataAccessException(1);
+            }
+            repository.deleteById(id);
+        } catch (EmptyResultDataAccessException e) {
+            throw new ResourceNotFoundException(id);
+        } catch (DataIntegrityViolationException e) {
+            throw new DatabaseException(e.getMessage());
+        }
+
     }
 
     public User update(Long id,User obj){
